issue #30 Favorited button doesn't always stick in large list
This bug was caused by UITableView recycling list item views and them not being refreshed after recycled.
When the item was favorited and scrolled out of view and then scrolled back in, the recycler used a list item which had not been favorited and the button was not updated.
The fix is to force a refresh of the view whenever it disappears/reappears.

Applied the same fix to the Favorites UITableview list tiems